修正DeparttmentsController的view型別錯誤bug

+++ WebApplication3/Models/Department.cs
@@ -22,7 +22,7 @@ namespace WebApplication3.Models
     
         public int DepartmentID { get; set; }
         public string Name { get; set; }
-        public short Budget { get; set; }
+        public decimal Budget { get; set; }
         public Nullable<System.DateTime> StartDate { get; set; }
         public Nullable<int> InstructorID { get; set; }
         public byte[] RowVersion { get; set; }
